+++ README.md
@@ -22,10 +22,14 @@
 
 0.1.0 - норм рефакторинг такой, осталось разбить на файлики и можно начинать делать новые ассисты
 
+- раскидать по файликам красиво
+
+0.1.1 - получается архитектура готова
+
 Что дальше
 ----
 
-- раскидать по файликам красиво
+- сделать ассист для док стринги
 - сделать юнит тесты
 - корректно обрабатывать ентры после \\ в любых строках
 - сделать проверку что мы в пу файле, сейчас ассист работает во всех, например в .txt
